requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.3.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py310he24745e_1
async-timeout=3.0.1=py_1000
attrs=22.1.0=pyh71513ae_1
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blinker=1.5=pyhd8ed1ab_0
brotlipy=0.7.0=py310h1961e1f_1004
bzip2=1.0.8=h0d85af4_4
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.9.24=h033912b_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310ha78151a_1
chardet=4.0.0=py310h2ec42d9_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=py310h2ec42d9_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.1=py310hf6deb26_0
diffusers=0.6.0=pyhd8ed1ab_0
discord.py=1.7.3=pyhd8ed1ab_0
filelock=3.8.0=pyhd8ed1ab_0
freetype=2.12.1=h3f81eb7_0
ftfy=6.1.1=pyhd8ed1ab_0
google-auth=2.13.0=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpc-cpp=1.48.1=h966d1d5_1
grpcio=1.48.1=py310h0f91fd3_1
huggingface_hub=0.10.1=pyhd8ed1ab_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.4=py310h2ec42d9_0
importlib_metadata=4.11.4=hd8ed1ab_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=hac89ed1_2
lcms2=2.12=h577c468_0
lerc=4.0.0=hb486fe8_0
libabseil=20220623.0=cxx17_h844d122_4
libblas=3.9.0=16_osx64_openblas
libcblas=3.9.0=16_osx64_openblas
libcxx=14.0.6=hccf4f1f_0
libdeflate=1.14=hb7f2c08_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=10_4_0_h97931a8_25
libgfortran5=11.3.0=h082f757_25
libiconv=1.17=hac89ed1_0
liblapack=3.9.0=16_osx64_openblas
libopenblas=0.3.21=openmp_h429af6e_3
libpng=1.6.38=ha978bb4_0
libprotobuf=3.21.7=hbc0c0cd_0
libsqlite=3.39.4=ha978bb4_0
libtiff=4.4.0=hdb44e8a_4
libwebp-base=1.2.4=h775f41a_0
libxcb=1.13=h0d85af4_1004
libzlib=1.2.13=hfd90126_4
llvm-openmp=14.0.4=ha654fa7_0
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py310h1961e1f_1
mkl=2022.1.0=h860c996_928
multidict=6.0.2=py310h1961e1f_1
ncurses=6.3=h96cf925_1
ninja=1.11.0=h1b54a9f_0
numpy=1.23.4=py310h1b7c290_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.0=h5d0d7b0_1
openssl=3.0.5=hfd90126_2
packaging=21.3=pyhd8ed1ab_0
pillow=9.2.0=py310h54af1cc_2
pip=22.3=pyhd8ed1ab_0
protobuf=4.21.7=py310h7a76584_0
pthread-stubs=0.4=hc929b4f_1001
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.6=hc14f532_0_cpython
python_abi=3.10=2_cp310
pytorch=1.12.1=cpu_py310h248013e_1
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py310h1961e1f_4
re2=2022.06.01=hb486fe8_0
readline=8.1.2=h3899abd_0
regex=2022.9.13=py310h90acd4f_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
sacremoses=0.0.53=pyhd8ed1ab_0
setuptools=65.5.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h6db0672_2
tbb=2021.6.0=hb8565cd_0
tensorboard=2.10.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py310h55d743e_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tk=8.6.12=h5dbffcc_0
tokenizers=0.11.4=py310h8776b5c_1
tqdm=4.64.1=pyhd8ed1ab_0
transformers=4.18.0=py310hecd8cb5_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022e=h191b570_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h35c211d_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
yaml=0.2.5=h0d85af4_2
yarl=1.7.2=py310h1961e1f_2
zipp=3.9.0=pyhd8ed1ab_0
zlib=1.2.13=hfd90126_4
zstd=1.5.2=hfa58983_4
